+++ HISTORY.rst
@@ -8,6 +8,11 @@
   directories.
 * Fixed duplicated FUSE directory removal.
 * Add link and view to show the parsed content of each document page.
+* Add a modelform for adding and editing transformation and perform
+  YAML validation of arguments.
+* Add stricted error checking to the crop transformation.
+* Update compressed files class module to work with Python 3.
+
 
 3.1.4 (2018-10-4)
 =================